+++ include/readout/ContinousLatencyBufferModel.hpp
@@ -16,7 +16,7 @@
 #include "ReadoutIssues.hpp"
 
 //#include <folly/ProducerConsumerQueue.h>
-#include "AccessableProducerConsumerQueue.hpp"
+#include "readout/AccessableProducerConsumerQueue.hpp"
 
 #include <memory>
 #include <utility>
