+++ include/readout/LatencyBufferConcept.hpp
