+++ include/readout/utils/FileSourceBuffer.hpp
@@ -8,7 +8,7 @@
 #ifndef READOUT_SRC_FILESOURCEBUFFER_HPP_
 #define READOUT_SRC_FILESOURCEBUFFER_HPP_
 
-#include "ReadoutIssues.hpp"
+#include "../ReadoutIssues.hpp"
 #include "readout/ReadoutLogging.hpp"
 
 #include "logging/Logging.hpp"
