+++ include/readout/utils/RateLimiter.hpp
@@ -8,7 +8,7 @@
 #ifndef READOUT_SRC_RATELIMITER_HPP_
 #define READOUT_SRC_RATELIMITER_HPP_
 
-#include "Time.hpp"
+#include "../Time.hpp"
 
 #include <atomic>
 #include <ctime>
