+++ plugins/DataRecorder.hpp
@@ -9,14 +9,14 @@
 #ifndef READOUT_PLUGINS_DATARECORDER_HPP_
 #define READOUT_PLUGINS_DATARECORDER_HPP_
 
-#include "BufferedFileWriter.hpp"
-#include "ReadoutStatistics.hpp"
-#include "RecorderConcept.hpp"
+#include "readout/utils/BufferedFileWriter.hpp"
+#include "readout/ReadoutStatistics.hpp"
+#include "readout/RecorderConcept.hpp"
 #include "appfwk/DAQModule.hpp"
 #include "appfwk/DAQSource.hpp"
 #include "appfwk/ThreadHelper.hpp"
-#include "readout/ReadoutTypes.hpp"
-#include "readout/ReusableThread.hpp"
+#include "readout/types/ReadoutTypes.hpp"
+#include "readout/utils/ReusableThread.hpp"
 #include "readout/datarecorder/Structs.hpp"
 
 #include <atomic>
