+++ plugins/DummyConsumer.cpp
@@ -11,7 +11,7 @@
 #include "appfwk/DAQModuleHelper.hpp"
 #include "readout/dummyconsumerinfo/InfoNljs.hpp"
 
-#include "ReadoutIssues.hpp"
+#include "readout/ReadoutIssues.hpp"
 #include "appfwk/cmd/Nljs.hpp"
 #include "logging/Logging.hpp"
 #include <string>
