+++ plugins/DummyConsumer.hpp
@@ -9,12 +9,12 @@
 #ifndef READOUT_PLUGINS_DUMMYCONSUMER_HPP_
 #define READOUT_PLUGINS_DUMMYCONSUMER_HPP_
 
-#include "ReadoutStatistics.hpp"
+#include "readout/ReadoutStatistics.hpp"
 #include "appfwk/DAQModule.hpp"
 #include "appfwk/DAQSource.hpp"
 #include "dfmessages/TimeSync.hpp"
-#include "readout/ReadoutTypes.hpp"
-#include "readout/ReusableThread.hpp"
+#include "readout/types/ReadoutTypes.hpp"
+#include "readout/utils/ReusableThread.hpp"
 
 #include <atomic>
 #include <fstream>
