+++ plugins/FakeCardReader.hpp
@@ -14,14 +14,14 @@
 #define READOUT_PLUGINS_FAKECARDREADER_HPP_
 
 // package
-#include "SourceEmulatorConcept.hpp"
-#include "readout/ReadoutTypes.hpp"
-#include "readout/ReusableThread.hpp"
+#include "readout/SourceEmulatorConcept.hpp"
+#include "readout/types/ReadoutTypes.hpp"
+#include "readout/utils/ReusableThread.hpp"
 #include "readout/fakecardreader/Structs.hpp"
 //#include "CreateSourceEmulator.hpp"
-#include "FileSourceBuffer.hpp"
-#include "RateLimiter.hpp"
-#include "ReadoutStatistics.hpp"
+#include "readout/utils/FileSourceBuffer.hpp"
+#include "readout/utils/RateLimiter.hpp"
+#include "readout/ReadoutStatistics.hpp"
 
 // appfwk
 #include "appfwk/DAQModule.hpp"
