+++ src/wib2/WIB2FrameProcessor.hpp
@@ -8,15 +8,15 @@
 #ifndef READOUT_SRC_WIB2FRAMEPROCESSOR_HPP_
 #define READOUT_SRC_WIB2FRAMEPROCESSOR_HPP_
 
-#include "ReadoutIssues.hpp"
-#include "ReadoutStatistics.hpp"
-#include "TaskRawDataProcessorModel.hpp"
-#include "Time.hpp"
+#include "../../include/readout/ReadoutIssues.hpp"
+#include "../../include/readout/ReadoutStatistics.hpp"
+#include "../../include/readout/TaskRawDataProcessorModel.hpp"
+#include "../../include/readout/Time.hpp"
 
 #include "dataformats/wib2/WIB2Frame.hpp"
 #include "logging/Logging.hpp"
 #include "readout/ReadoutLogging.hpp"
-#include "readout/ReadoutTypes.hpp"
+#include "readout/types/ReadoutTypes.hpp"
 
 #include <atomic>
 #include <functional>
