+++ test/apps/test_bufferedfilereader_app.cxx
@@ -6,10 +6,10 @@
  * Licensing/copyright details are in the COPYING file that you should have
  * received with this code.
  */
-#include "BufferedFileReader.hpp"
+#include "readout/utils/BufferedFileReader.hpp"
 
 #include "logging/Logging.hpp"
-#include "readout/ReadoutTypes.hpp"
+#include "readout/types/ReadoutTypes.hpp"
 
 #include <atomic>
 #include <chrono>
