+++ test/apps/test_ratelimiter_app.cxx
@@ -6,8 +6,8 @@
  * Licensing/copyright details are in the COPYING file that you should have
  * received with this code.
  */
-#include "RandomEngine.hpp"
-#include "RateLimiter.hpp"
+#include "readout/utils/RandomEngine.hpp"
+#include "readout/utils/RateLimiter.hpp"
 
 #include "logging/Logging.hpp"
