+++ test/apps/test_skiplist_app.cxx
@@ -6,12 +6,12 @@
  * Licensing/copyright details are in the COPYING file that you should have
  * received with this code.
  */
-#include "RandomEngine.hpp"
-#include "RateLimiter.hpp"
+#include "readout/utils/RandomEngine.hpp"
+#include "readout/utils/RateLimiter.hpp"
 
 #include "logging/Logging.hpp"
 
-#include "readout/ReadoutTypes.hpp"
+#include "readout/types/ReadoutTypes.hpp"
 
 #include "folly/ConcurrentSkipList.h"
