+++ unittest/BufferedReadWrite_test.cxx
@@ -13,10 +13,10 @@
 
 #include "boost/test/unit_test.hpp"
 
-#include "BufferedFileReader.hpp"
-#include "BufferedFileWriter.hpp"
+#include "readout/utils/BufferedFileReader.hpp"
+#include "readout/utils/BufferedFileWriter.hpp"
 #include "logging/Logging.hpp"
-#include "readout/ReadoutTypes.hpp"
+#include "readout/types/ReadoutTypes.hpp"
 
 #include <cstdio>
 #include <string>
